+++ CHANGELOG.md
@@ -76,6 +76,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added a new named data mix that we used for the 32B run
 - Added internal OLMo3 7B midtraining and long-context configs.
 - Added ability to convert OLMo3 models to/from HF format with support for rope scaling configs.
+- Added the `WSDS` (Warmup-Stable-Decay-Simplified) learning rate scheduler.
 - Added a script that can pull out a single training batch from a training job
 
 ## [v2.2.0](https://github.com/allenai/OLMo-core/releases/tag/v2.2.0) - 2025-08-26